[IMP] shipment_advice_planner_toursolver: Recreate toursolver task
It may happen that the shipment advice is created after the toursolver task is
completed, and the user decides to add a picking to it.
It's easier to have an action on the shipment advice to recreate the toursolver
task without going through the planner again.

As a safeguard, we do not allow this action unless the previous task has been
removed. When the task is recreated, the new task will not generate a new
shipment advice but will be assigned to the existing one.